requirements.txt
requirements.txt
streamlit
streamlit-folium
folium
pillow
pandas
